requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
asttokens==3.0.1
attrs==25.4.0
backoff==2.2.1
certifi==2025.11.12
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
dataclasses-json==0.6.7
datasets==4.4.2
debugpy==1.8.19
decorator==5.2.1
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring-parser==0.17.0
executing==2.2.1
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.10.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.2.3
hyperframe==6.1.0
identify==2.6.15
idna==3.11
importlib-metadata==8.7.1
instructor==1.13.0
ipykernel==7.1.0
ipython==9.8.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jieba==0.42.1
jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==8.7.0
jupyter-core==5.9.1
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.5
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
langfuse==3.11.1
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.0
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
multidict==6.7.0
multiprocess==0.70.18
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.0
openai==2.14.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
portalocker==3.2.0
pre-commit==4.5.1
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==22.0.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
qdrant-client==1.16.2
ragas==0.4.1
rank-bm25==0.2.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
scikit-network==0.33.5
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.45
stack-data==0.6.3
tenacity==9.1.2
tiktoken==0.12.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
ty==0.0.5
typer==0.20.1
typer-slim==0.20.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.2
uuid-utils==0.12.0
virtualenv==20.35.4
wcwidth==0.2.14
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20
gunicorn==23.0.0
